refresh token check
While validating a refresh token,
fetch saved refresh token from DB
and comare the two versions.
If they do not match (either because nothing
was found in DB or because not the last version
of the token is being validated), do not validate the token.
That way, refresh token may be revoked
by deleting the last version from DB.